docs: add developer docs, playlist_filters module
Why?
Explanation about what aspects of djtools are extensible is necessary
for developers to contribute.

What?
Separate PlaylistFilters into their own module. Add several docs for
developers.
