+++ docs/tutorials/developer_docs/index.md
@@ -1,0 +1,6 @@
+# Developer Documentation
+
+* [System Overview](system_overview.md)
+* [Adding DJ Software Support to the Collections Package](new_collections.md)
+* [Using the Collections API](collections_api.md)
+* [Adding Custom PlaylistFilters](playlist_filters.md)
